# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dating Project",
        "description": "This is a sample Dating API server.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/information": {
            "get": {
                "tags": [
                    "Dictionaries"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\DictionaryController::information",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Personal info"
                    }
                }
            }
        },
...
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\UserController::user",
                "responses": {
                    "200": {
                        "description": "User data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string"
                                        },
                                        "data": {
                                            "type": "object",
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\UserController::update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "firstname": {
                                        "type": "string"
                                    },
                                    "middlename": {
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string"
                                        },
                                        "data": {
                                            "type": "object",
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/information": {
            "put": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\UserController::updateInformation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "relationship_id": {
                                        "type": "integer"
                                    },
                                    "living_id": {
                                        "type": "integer"
                                    },
                                    "children_id": {
                                        "type": "integer"
                                    },
                                    "smoking_id": {
                                        "type": "integer"
                                    },
                                    "drinking_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string"
                                        },
                                        "data": {
                                            "type": "object",
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user/filters": {
            "put": {
                "tags": [
                    "User"
                ],
                "operationId": "App\\Http\\Controllers\\Api\\UserController::updateFilters",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "gender": {
                                        "type": "string"
                                    },
                                    "age_from": {
                                        "type": "integer"
                                    },
                                    "age_to": {
                                        "type": "integer"
                                    },
                                    "city_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string"
                                        },
                                        "data": {
                                            "type": "object",
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/User"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "UserFilter": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "format": "integer"
                            },
                            "reason": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserSearchReason"
                                    }
                                ]
                            },
                            "gender": {
                                "type": "string"
                            },
                            "age_from": {
                                "type": "integer"
                            },
                            "age_to": {
                                "type": "integer"
                            },
                            "city_id": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "UserInfoOption": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "format": "integer"
                            },
                            "name": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "UserInformation": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "relationship": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserInfoOption"
                                    }
                                ]
                            },
                            "living": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserInfoOption"
                                    }
                                ]
                            },
                            "children": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserInfoOption"
                                    }
                                ]
                            },
                            "smoking": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserInfoOption"
                                    }
                                ]
                            },
                            "drinking": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserInfoOption"
                                    }
                                ]
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "UserOptions": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "gender": {
                                "type": "string",
                                "format": "string"
                            },
                            "birthday": {
                                "type": "string",
                                "format": "date"
                            },
                            "country_id": {
                                "type": "integer",
                                "format": "integer"
                            },
                            "region_id": {
                                "type": "integer",
                                "format": "integer"
                            },
                            "city_id": {
                                "type": "integer",
                                "format": "integer"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "UserSearchReason": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "format": "integer"
                            },
                            "name": {
                                "type": "string",
                                "format": "string"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "User": {
                "type": "object",
                "allOf": [
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "format": "integer"
                            },
                            "middlename": {
                                "type": "string",
                                "format": "string"
                            },
                            "lastname": {
                                "type": "string",
                                "format": "string"
                            },
                            "avatar_url": {
                                "type": "string",
                                "format": "string"
                            },
                            "last_activity": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "options": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserOptions"
                                    }
                                ]
                            },
                            "information": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserInformation"
                                    }
                                ]
                            },
                            "filters": {
                                "type": "object",
                                "allOf": [
                                    {
                                        "$ref": "#/components/schemas/UserFilter"
                                    }
                                ]
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        }
    }
}